# Swagger Mock & Documentation/swagger-client-generated/swagger.json
{
  "swagger" : "2.0",
  "info" : {
    "description" : "",
    "version" : "1.0.0",
    "title" : "API Hotel"
  },
  "host" : "virtserver.swaggerhub.com",
  "basePath" : "/MyBookingServices/DevOps_Cloud/1.0.0",
...
  "schemes" : [ "https" ],
  "paths" : {
    "/user/register" : {
      "post" : {
        "tags" : [ "User" ],
        "summary" : "Creates a new user account.",
        "description" : "Provided email address must be unique.",
        "parameters" : [ {
          "in" : "body",
          "name" : "userItem",
          "description" : "User item to add",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/UserRegister"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "",
            "schema" : {
              "$ref" : "#/definitions/inline_response_200"
            }
          },
          "400" : {
            "description" : "",
            "schema" : {
              "$ref" : "#/definitions/inline_response_400"
            }
          },
          "409" : {
            "description" : "",
            "schema" : {
              "$ref" : "#/definitions/inline_response_409"
            }
          }
        }
      }
    },
    "/user/auth" : {
      "post" : {
        "tags" : [ "User" ],
        "summary" : "User authentification.",
        "description" : "Provided email address must be unique.",
        "parameters" : [ {
          "in" : "body",
          "name" : "userAuth",
          "description" : "User auth",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/UserAuth"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "",
            "schema" : {
              "$ref" : "#/definitions/inline_response_200_1"
            }
          },
          "400" : {
            "description" : "",
            "schema" : {
              "$ref" : "#/definitions/inline_response_400"
            }
          },
          "401" : {
            "description" : "",
            "schema" : {
              "$ref" : "#/definitions/inline_response_401"
            }
          }
        }
      }
    },
    "/user/" : {
      "get" : {
        "tags" : [ "User" ],
        "summary" : "Get Users.",
        "description" : "Display All Users.",
        "parameters" : [ ],
        "responses" : {
          "200" : {
            "description" : "",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/inline_response_200_2"
              }
            }
          },
          "401" : {
            "description" : "",
            "schema" : {
              "$ref" : "#/unauthorize/responses/UnauthorizedError"
            }
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ]
      }
    },
...
    "/user/logout" : {
      "delete" : {
        "tags" : [ "User" ],
        "summary" : "Logout user.",
        "description" : "",
        "parameters" : [ ],
        "responses" : {
          "200" : {
            "description" : "",
            "schema" : {
              "$ref" : "#/definitions/inline_response_200_5"
            }
          },
          "401" : {
            "description" : "",
            "schema" : {
              "$ref" : "#/unauthorize/responses/UnauthorizedError"
            }
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ]
      }
    },
...
  },
  "definitions" : {
    "UserRegister" : {
      "type" : "object",
      "required" : [ "email", "password" ],
      "properties" : {
        "email" : {
          "type" : "string",
          "example" : "swagger@hotmail.com",
          "description" : "Email"
        },
        "password" : {
          "type" : "string",
          "example" : "124K8t9iHn!",
          "description" : "Password"
        },
        "nom" : {
          "type" : "string",
          "example" : "Test",
          "description" : "Name"
        }
      }
    },
    "UserAuth" : {
      "type" : "object",
      "required" : [ "email", "password" ],
      "properties" : {
        "email" : {
          "type" : "string",
          "example" : "swagger@hotmail.com",
          "description" : "Email"
        },
        "password" : {
          "type" : "string",
          "example" : "124K8t9iHn!",
          "description" : "Password"
        }
      }
    },
...
    "inline_response_200" : {
      "type" : "object",
      "properties" : {
        "message" : {
          "type" : "string",
          "example" : "User created successfully!",
          "description" : "Message"
        },
        "ok" : {
          "type" : "boolean",
          "example" : true,
          "description" : ""
        }
      }
    },
    "inline_response_400" : {
      "type" : "object",
      "properties" : {
        "message" : {
          "type" : "string",
          "example" : "Bad request parameters: {}",
          "description" : "Message"
        },
        "ok" : {
          "type" : "boolean",
          "example" : false,
          "description" : ""
        }
      }
    },
    "inline_response_409" : {
      "type" : "object",
      "properties" : {
        "message" : {
          "type" : "string",
          "example" : "User already registered",
          "description" : "Error"
        },
        "ok" : {
          "type" : "boolean",
          "example" : false,
          "description" : ""
        }
      }
    },
    "inline_response_200_1_data" : {
      "type" : "object",
      "properties" : {
        "email" : {
          "type" : "string",
          "example" : "swagger@hotmail.com",
          "description" : "Email"
        },
        "nom" : {
          "type" : "string",
          "example" : "Test",
          "description" : "Name"
        },
        "token" : {
          "type" : "string",
          "example" : "eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJpYXQiOjE1OTU4ODQxNDIsIm5iZiI6MTU5NTg4NDE0MiwianRpIjoiNzQ4YTc4NGMtY2YwYi00MjA2LWI4NjItODFkM2JlMjllYzQxIiwiZXhwIjoxNTk1OTcwNTQyLCJpZGVudGl0eSI6eyJlbWFpbCI6InRoaXJ1Y190QGV0bmEtYWx0ZXJuYW5jZS5uZXQiLCJwYXNzd29yZCI6InBhc3N3b3JkIn0sImZyZXNoIjpmYWxzZSwidHlwZSI6ImFjY2VzcyJ9.F112-shUH2pl5ejOhlVupoI8bU1plQi6B4Xs_yB_fuE",
          "description" : "Token"
        }
      }
    },
    "inline_response_200_1" : {
      "type" : "object",
      "properties" : {
        "data" : {
          "$ref" : "#/definitions/inline_response_200_1_data"
        },
        "ok" : {
          "type" : "boolean",
          "example" : true,
          "description" : ""
        }
      }
    },
    "inline_response_401" : {
      "type" : "object",
      "properties" : {
        "message" : {
          "type" : "string",
          "example" : "invalid email or password",
          "description" : "Error"
        },
        "ok" : {
          "type" : "boolean",
          "example" : false,
          "description" : ""
        }
      }
    },
    "user_id" : {
      "type" : "object",
      "properties" : {
        "$iod" : {
          "type" : "string",
          "example" : "5f1f4267af0a8a1cc488a4d3",
          "description" : ""
        }
      }
    },
    "user_password" : {
      "type" : "object",
      "properties" : {
        "$binary" : {
          "type" : "string",
          "example" : "JDJiJDEyJC51OWx5VUZuV3ouTElRM3k3d1VBbE94d0U0MDYzUTFhN2lKeEJHNUdlNmN4dG1BWDVpUS9X",
          "description" : ""
        },
        "$type" : {
          "type" : "string",
          "example" : "00",
          "description" : ""
        }
      }
    },
    "inline_response_200_2" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "$ref" : "#/definitions/user_id"
        },
        "email" : {
          "type" : "string",
          "example" : "swagger@hotmail.com",
          "description" : "Email"
        },
        "nom" : {
          "type" : "string",
          "example" : "Test",
          "description" : "Name"
        },
        "password" : {
          "$ref" : "#/definitions/user_password"
        }
      }
    },
...
    "inline_response_200_5" : {
      "type" : "object",
      "properties" : {
        "message" : {
          "type" : "string",
          "example" : "Successfully logged out. Please reconnect",
          "description" : "Message"
        }
      }
    },
...
  }
}